backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
channels==4.0.0
charset-normalizer==3.3.0
Django==4.2.6
django-nextjs==2.4.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
sqlparse==0.4.4
websockets==11.0.3
yarl==1.9.2
black
djangorestframework
django-cors-headers
embedchain